+++ _docs/202603/202603182212_sumber-glosarium-pariwisata.sql
@@ -1,0 +1,23 @@
+-- Menambahkan sumber kamus pariwisata, perhotelan, dan usaha perjalanan wisata
+-- dari Badan Pengembangan (dan Pembinaan) Bahasa
+
+INSERT INTO sumber (kode, nama, glosarium, keterangan)
+VALUES
+  (
+    'KPRW',
+    'Kamus Pariwisata',
+    true,
+    'Arini, D.E., Warganegara, H., Simanungkalit, D.C., Pamuji, G.M., Miharti, T., & Rahmawati, D. (2018). *[Kamus Pariwisata](https://badanbahasa.kemendikdasmen.go.id/resource/doc/files/Kamus_Pariwisata.pdf)*. Badan Pengembangan dan Pembinaan Bahasa.'
+  ),
+  (
+    'KHOT',
+    'Kamus Perhotelan',
+    true,
+    'Rahmawati, H. (2019). *[Kamus Perhotelan](https://badanbahasa.kemendikdasmen.go.id/resource/doc/files/Kamus_Perhotelan.pdf)*. Badan Pengembangan Bahasa dan Perbukuan.'
+  ),
+  (
+    'KUPW',
+    'Kamus Usaha Perjalanan Wisata',
+    true,
+    'Pamuji, G.M. (2019). *[Kamus Usaha Perjalanan Wisata](https://badanbahasa.kemendikdasmen.go.id/resource/doc/files/Kamus_UPW.pdf)*. Badan Pengembangan Bahasa dan Perbukuan.'
+  );